🔥 Trim Dev-screen leftovers from main process
- Drop IPC.permissionHas and IPC.captureOnce end-to-end (channel,
  handler, preload bridge, SlowblinkAPI member) — neither was reached
  by the renderer after the Dev screen moved out
- Relocate LogEntry/LogLevel/LOG_BUFFER_SIZE into src/main/logger.ts
  and OverviewDebug into src/main/overview/debug.ts; they no longer
  cross the preload bridge, so shared/types.ts shouldn't own them
- Drop the stale `export { getOverview }` re-export from overview/debug
