+++ src/shared/ipc-channels.ts
@@ -7,10 +7,8 @@ export const IPC = {
   statusGet: 'status:get',
   capturePause: 'capture:pause',
   captureResume: 'capture:resume',
-  captureOnce: 'capture:once',
   permissionRequest: 'permission:request',
   permissionOpen: 'permission:open',
-  permissionHas: 'permission:has',
   permissionAccessibilityRequest: 'permission:accessibility:request',
   permissionAccessibilityOpen: 'permission:accessibility:open',
   dataDeleteAll: 'data:deleteAll',
